// frontend/src/module/image-mng/image-mng/model/img-mng.model.ts
export class Image {
    id:string;
    name: string;
    type: string;
    os: string;
    bits: string;
    createTime: string;
    status: string;
    progress: string;
    imageOwner:string;
    areaId:string;
    areaName: string;
    creatorName: string;
    creatorId: string;
    orgId: string;
    orgName:string;
    description: string;
    selected: boolean = false;
    nameEditing: boolean;
    desEditing:boolean;
    constructor() {
    }
}